requirements.txt
streamlit
google-generativeai
python.env
pdf2image 
